utilities
some small changes that helps the usage.
The `domRuler.track` has no use when the element is not attached to the dom. In fact it is horrible because it only sends empty rectangles. Adding a non-breaking param to aways watch the layout resizing helps that.
The material-tab used to notify its visibility once it is activated, however, the content is not ready yet. Moving it from the activation methods to the execution of the `ngIf` fixes that.